+++ commands/src/main/java/net/apartium/cocoabeans/commands/RegisteredCommandVariant.java
@@ -10,16 +10,20 @@
 
 package net.apartium.cocoabeans.commands;
 
+import net.apartium.cocoabeans.commands.parsers.ArgumentParser;
 import net.apartium.cocoabeans.commands.requirements.ArgumentRequirement;
+import net.apartium.cocoabeans.commands.requirements.Requirement;
 
 import java.lang.invoke.MethodHandle;
 import java.lang.reflect.Parameter;
 import java.lang.reflect.Type;
+import java.util.List;
 
 public record RegisteredCommandVariant(
         MethodHandle method,
         Parameter[] parameters,
         CommandNode commandNode,
+        List<ArgumentIndex<?>> argumentIndexList,
         int priority
 ) {