+++ commands/src/main/java/net/apartium/cocoabeans/commands/requirements/Requirement.java
@@ -10,6 +10,9 @@
 
 package net.apartium.cocoabeans.commands.requirements;
 
+import java.util.Collections;
+import java.util.List;
+
 /**
  * Represents a requirement to run a command or a sub command, eg a permission, sender type (player / console etc)
  * @see RequirementFactory
@@ -23,6 +26,15 @@ public interface Requirement {
      * @return if sender meets requirement or else given an error
      */
     RequirementResult meetsRequirement(RequirementEvaluationContext context);
+
+    /**
+     * Get list of classes that will be return if requirement is met
+     * Note: By default returns an empty list
+     * @return list of classes that will be return if requirement is met
+     */
+    default List<Class<?>> getTypes() {
+        return Collections.emptyList();
+    }
 
 
 }