+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - [commands] WithParser support keyword change
 - [commands] Add new constructor to every `ArgumentParser` with those two arguments `int, String`
 - [commands] Add QuotedStringParser & ParagraphParser
+- [commands] Improve ObjectMapper to pre calculate all possible parsers
 
 ## 0.0.35
 - [spigot] **hotfix** Wrong package for `PlayerVisibilityController_1_8_R1`